split code into diagnostics.rs
